checker/requirements.txt
requests
psycopg2-binary
dj-database-url